Revert "modules: carbonrom: {File, Rom}: Timestamp: Convert to int"
time.Unix use int64 timestamp

This reverts commit 0705fad1b8674c62ce8ddd87e01b5de4b80c3aa7.

+++ modules/carbonrom/carbonrom.go
@@ -10,13 +10,13 @@ import (
 
 type File struct {
 	Filename string
-	Timestamp int
+	Timestamp int64
 }
 
 type Rom struct {
 	RomPath   string
 	Md5Url    string
-	Timestamp int
+	Timestamp int64
 }
 
 func (r Rom) RomUrl() string {
